Feat design revamp

+++ ip-atlas/routes/atlas.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, render_template, request, abort
+from flask import Blueprint, render_template, request, abort, jsonify
 from jinja2 import TemplateNotFound
 from helper import *
 
@@ -16,6 +16,11 @@ def list():
     # printJson()
     data = loadJson()
     return render_template('ip/list.html', data=data)
+
+@bp_atlas.route('/ip/ping/<ip_address>')
+def ping_ip(ip_address):
+    pingable = isIpPingable(ip_address)
+    return jsonify({'pingable': pingable})
 
 @bp_atlas.route('/ip/<id>')
 def show(id):
@@ -38,17 +43,23 @@ def save():
     if request.method == 'POST':
         # get form data
         name = request.form.get('name')
+        tags = request.form.get('tags')
         ipv4 = request.form.get('ipv4')
         ipv6 = request.form.get('ipv6')
         ports = request.form.get('ports')
+        tags = request.form.get('tags')
 
         if ports:
             # convert ports from comma separated to list
             ports = ports.split(',')
+        if tags:
+            # convert tags from comma separated to list
+            tags = tags.split(',')
         
         #print(name, ipv4, ipv6, ports)
-        writeJson(name, ipv4, ipv6, ports)
-    return render_template('ip/list.html')
+        writeJson(name, tags, ipv4, ipv6, ports)
+        data = loadJson()
+    return render_template('ip/list.html', data=data)
 
 
 
@@ -56,6 +67,29 @@ def save():
 @bp_atlas.route('/ip/delete/<id>')
 def delete(id):
     id = int(id)
-    deleteHost(id)
-    print("Host with id: ", id, " deleted")
-    return render_template('ip/list.html')
+    confirmed = request.args.get('confirmed')
+    data = loadJson()
+    
+    if confirmed == 'true':
+        deleteHost(id)
+        # print("Host with id: ", id, " deleted")
+        return render_template('ip/list.html', data=data)
+    else:
+        return render_template('ip/list.html', data=data)
+
+# confirm delete host
+@bp_atlas.route('/ip/confirmdel/<id>')
+def confirmdel(id):
+    id = int(id)
+    data = getHostById(id)
+    return render_template('ip/confirmdel.html', data=data)
+
+
+@bp_atlas.route('/port/list')
+def port():
+    
+    return render_template('port/list.html')
+
+@bp_atlas.route('/statistic')
+def statistic():
+    return render_template('statistic.html')
